Blake3 (#1)
* digest: promote blake3 to first-class digest

The dual module approach for blake3 was slightly awkward. Since it
provides similar usability with a massive bump in performance, it's
extremely likely to land as a registered algorithm in the image-spec.

This PR removes the secondary module, which made it difficult to test as
a unit. This may break users who are using HEAD versions of the package.
For a new release, this will be backwards compatible. The other drawback
is that the zeebo/blake3 will now be a dependency but this can be
replaced transparently by the standard libary in the future.

In addition to promoting blake3, this makes a few style adjustments to
be in line with Go's style guidelines.

Signed-off-by: Stephen Day <stephen.day@getcruise.com>

* fix: update stevvooe's blake3 PR

https://github.com/opencontainers/go-digest/pull/66
https://github.com/opencontainers/image-spec/pull/1240
Signed-off-by: Ramkumar Chinchani <rchincha.dev@gmail.com>

* fix: add a length test for blake3

Signed-off-by: Ramkumar Chinchani <rchincha.dev@gmail.com>

* fix: add a Makefile

make
make build
make test

Signed-off-by: Ramkumar Chinchani <rchincha.dev@gmail.com>

* fix: merge conflicts

Signed-off-by: Ramkumar Chinchani <rchincha.dev@gmail.com>

* fix: blake3 pulls in golang 1.22.x dep

2025-04-25T19:33:20.2495501Z go: downloading github.com/zeebo/blake3 v0.2.4
2025-04-25T19:33:20.3154128Z go: downloading github.com/klauspost/cpuid/v2 v2.2.10
2025-04-25T19:33:20.3959610Z github.com/klauspost/cpuid/v2: cannot compile Go 1.22 code
2025-04-25T19:33:21.1059807Z FAIL github.com/opencontainers/go-digest [build failed]
2025-04-25T19:33:21.1060449Z FAIL github.com/opencontainers/go-digest/digestset [build failed]
2025-04-25T19:33:21.1219422Z ##[error]Process completed with exit code 1.
Signed-off-by: Ramkumar Chinchani <rchincha.dev@gmail.com>

---------

Signed-off-by: Stephen Day <stephen.day@getcruise.com>
Signed-off-by: Ramkumar Chinchani <rchincha.dev@gmail.com>
Co-authored-by: Stephen Day <stephen.day@getcruise.com>